+++ warehouse/products/helpers/importers/importer.py
@@ -1,10 +1,7 @@
 from typing import NamedTuple
 
-import pandas
-
 from warehouse.products import model
 from warehouse.suppliers import model as supplier_model
-from warehouse.common.helpers import BaseFactory
 
 
 class IncompleteImporterMapping(KeyError):
@@ -14,82 +11,6 @@ class IncompleteImporterMapping(KeyError):
 class ProductPair(NamedTuple):
     parsed_product: dict
     supplier_product: supplier_model.Supplierproduct
-
-
-class StockParser:
-    def __init__(self, file_path, columns, normalize_columns):
-        """Attempts to find the columns and values from a passed html file object.
-
-        Args:
-            file_path (StringIO): The StringIO object containing the HTML with the table.
-            columns (tuple[str]): The columns that we are interested in
-            normalize_columns (tuple[str]): The column containing the product name that should be normalized to match the database value.
-        """
-        self.columns = columns
-        self.normalize_columns = normalize_columns
-        self.file_path = file_path
-
-    def Parse(self):
-        """Start the parsing process."""
-        dataframes = self._parse()
-        return self._process_dataframes(dataframes)
-
-    def _parse(self):
-        """Read the file and attempt to find the columns."""
-        return pandas.read_html(self.file_path, header=0)
-
-    def _process_dataframes(self, dataframes):
-        """Process the list of dataframes containing table elements.
-
-        Args:
-            dataframes (list[DataFrame]): List with dataframes found by pandas.read_html
-
-        Returns:
-            list[dict]: Returns the list with the matches.
-        """
-        # Because multiple tables can be present in a page we can have multiple dataframes.
-        return [
-            self._process_dataframe(dataframe.to_dict("record"))
-            for dataframe in dataframes
-        ]
-
-    def _process_dataframe(self, dataframe):
-        """Process the dataframe by normalizing the values contained in the columns.
-
-        Returns:
-            list[dict]: A list of dictionaries with the processed matches.
-        """
-        results = []
-
-        for result in dataframe:
-            if any(
-                missing_columns := [
-                    column for column in self.columns if column not in result.keys()
-                ]
-            ):
-                raise KeyError(
-                    f"The following columns could not be found: {missing_columns}"
-                )
-
-            results.append(self._normalize(result))
-
-        return results
-
-    def _normalize(self, result):
-        """Normalize only the columns which are of interest. This should be the column containing the product name."""
-        clean_result = self._remove_unwanted_keys(result)
-
-        for column in self.normalize_columns:
-            clean_result[column] = clean_result[column].replace("/", "_")
-        return clean_result
-
-    def _remove_unwanted_keys(self, result):
-        """Create a copy of the result and mutate the object by removing keys that are not sought after."""
-        copy = dict(result)
-        for key in result.keys():
-            if key not in self.columns:
-                del copy[key]
-        return copy
 
 
 class StockImporter:
@@ -202,12 +123,6 @@ class StockImporter:
         return product.Refresh()
 
 
-def csv_parser(file, interested_columns: tuple):
-    data = pandas.read_csv(file, skip_blank_lines=True, usecols=interested_columns)
-    data.dropna(how="all", inplace=True)
-    return data.to_dict("records")
-
-
 class CsvImporter(StockImporter):
     def Import(
         self, parsed_results: list[dict], products: list[model.Product]
@@ -224,16 +139,3 @@ class CsvImporter(StockImporter):
         self.products = list(products)
         self._import_parsed_results(self.parsed_results)
         return self._processed_products, self._unprocessed_products
-
-
-class ParserFactory(BaseFactory):
-    """Factory class to keep track of all supported parsers"""
-
-    def register_base_classes(self):
-        self.register("html_table", StockParser)
-        self.register("csv", csv_parser)
-
-
-if __name__ == "__main__":
-    x = csv_parser("/home/stef/Downloads/pricelist_short.csv", ("article_number",))
-    print(x)
